moving tools.cmake
