+++ conan/tools/cmake/__init__.py
@@ -1,0 +1,3 @@
+# noinspection PyUnresolvedReferences
+from .cmake import CMake
+from .toolchain import CMakeToolchain
