+++ conan/tools/cmake/android.py
