+++ conans/test/functional/build_helpers/cmake_test.py
@@ -7,7 +7,7 @@ from conans.test.utils.tools import TestClient
 
 
 @pytest.mark.tool_cmake
-class CMakeBuildHelper(unittest.TestCase):
+class CMakeBuildHelperTest(unittest.TestCase):
     def test_get_version_no_toolchain(self):
         client = TestClient()
         conanfile = textwrap.dedent("""
